Move serialization logic to its own file
